backend/requirements.txt
fastapi==0.115.5
pydantic==2.10.1
pydantic-settings==2.6.1
pyzbar==0.1.9
requests==2.32.3
streamlit==1.40.2
opencv-python==4.10.0.84
python-dotenv==1.0.1
uvicorn[standard]==0.32.1
python-multipart==0.0.12
mysql-connector-python==8.3.0
pandas==2.2.2
reportlab==4.2.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

